Fix linter files
