+++ my_list.rb
@@ -1,0 +1,23 @@
+require_relative './my_enumerable'
+class MyList
+  include MyEnumerable
+  def initialize(*list)
+    @list = list
+  end
+
+  def each(&all_elements)
+    @list.each(&all_elements)
+  end
+end
+list = MyList.new(1, 2, 3, 4)
+# test all_method
+p(list.all? { |e| e < 5 })
+# true
+p(list.all? { |e| e > 5 })
+# false
+p(list.any? { |e| e == 2 })
+# true
+p(list.any? { |e| e == 5 })
+# false
+p(list.filter(&:even?))
+# [2, 4]
